requirements.txt
pymongo>=4.8.0
dnspython>=2.3.0
Flask>=3.0.0
python-dotenv>=1.0.0
Faker>=19.0.0
certifi